refactor: 重构electron主窗口服务
